Maybe remove MergedCollider
