+++ zzre.core/collider/WorldCollider.cs
@@ -3,163 +3,193 @@ using System.Collections.Generic;
 using System.Linq;
 using zzio.rwbs;
 using zzio;
+using System.Diagnostics;
+using System.Numerics;
+using System;
 
 namespace zzre;
 
-public readonly record struct WorldTriangleId(int AtomicIdx, int TriangleIdx);
+/*
+ * For worlds we merge the atomic sections and the underlying RWCollisions together
+ * This brings a bit more performance and a whole lot less complexity at runtime
+ * Unfortunately also some amount of memory
+ * We still have a sub-type to be able to query atomic sections and triangles by id
+ */
 
-public readonly record struct WorldTriangleInfo(RWAtomicSection Atomic, VertexTriangle VertexTriangle)
+public sealed class WorldCollider : TreeCollider
 {
-    public Triangle Triangle => new(
-        Atomic.vertices[VertexTriangle.v1],
-        Atomic.vertices[VertexTriangle.v2],
-        Atomic.vertices[VertexTriangle.v3]);
-}
-
-public sealed class WorldCollider : BaseGeometryCollider
-{
-    public RWWorld World { get; }
-    public Box Box { get; }
-
     private readonly RWAtomicSection[] atomicSections;
-    private readonly Section rootSection;
-    private readonly IReadOnlyDictionary<RWAtomicSection, AtomicTreeCollider> atomicColliders;
-    protected override IRaycastable CoarseCastable => Box;
-    protected override IIntersectable CoarseIntersectable => Box;
 
-    public WorldCollider(RWWorld world)
+    private WorldCollider(
+        Box coarse, RWCollision collision,
+        ReadOnlyMemory<Triangle> triangles,
+        ReadOnlyMemory<WorldTriangleId> triangleIds,
+        RWAtomicSection[] atomicSections)
+        : base(coarse, collision, location: null, triangles, triangleIds)
     {
-        World = world;
+        this.atomicSections = atomicSections;
+    }
 
-        atomicSections = World
+    public static WorldCollider Create(RWWorld world)
+    {
+        var atomicSections = world
             .FindAllChildrenById(SectionId.AtomicSection, recursive: true)
             .Cast<RWAtomicSection>()
             .ToArray();
-        var colliders = atomicSections
-            .Select((section, i) => new AtomicTreeCollider(section, i))
+        var atomicCollisions = atomicSections
+            .Select(section => section.FindChildById(SectionId.CollisionPLG, true) ??
+                                CreateNaiveCollision(section.triangles.Length))
+            .Cast<RWCollision>()
             .ToArray();
 
-        atomicColliders = colliders.ToDictionary(c => c.Atomic, c => c);
-        Box = colliders.Aggregate(colliders.First().Box, (box, atomic) => box.Union(atomic.Box));
+        // figure out AABB
+        var boxMin = atomicSections.Aggregate(Vector3.One * float.PositiveInfinity,
+            (acc, section) => Vector3.Min(acc, Vector3.Min(section.bbox1, section.bbox2)));
+        var boxMax = atomicSections.Aggregate(Vector3.One * float.NegativeInfinity,
+            (acc, section) => Vector3.Max(acc, Vector3.Max(section.bbox1, section.bbox2)));
+        var box = Box.FromMinMax(boxMin, boxMax);
 
-        var rootPlane = World.FindChildById(SectionId.PlaneSection, false);
-        var rootAtomic = World.FindChildById(SectionId.AtomicSection, false);
-        rootSection = rootPlane ?? rootAtomic ?? throw new InvalidDataException("RWWorld has no geometry");
-        if (rootPlane != null && rootAtomic != null)
-            throw new InvalidDataException("RWWorld has both a root plane and a root atomic");
+        // unmap triangles and remove degenerated ones
+        var triangleCount = atomicCollisions.Sum(s => s.map.Length);
+        var triangles = new Triangle[triangleCount];
+        var triangleIds = new WorldTriangleId[triangleCount];
+        var baseMapIndices = new int[atomicSections.Length];
+        var triangleCounts = new int[atomicSections.Length]; // that is original count but without the degenerated ones
+        var map = new int[triangleCount];
+        var triangleI = 0;
+        for (int sectionI = 0; sectionI < atomicSections.Length; sectionI++)
+        {
+            var section = atomicCollisions[sectionI];
+            var localTriangles = atomicSections[sectionI].triangles;
+            var vertices = atomicSections[sectionI].vertices;
+            baseMapIndices[sectionI] = triangleI;
+            for (int localTriangleI = 0; localTriangleI < section.map.Length; localTriangleI++)
+            {
+                var t = localTriangles[section.map[localTriangleI]];
+                triangles[triangleI] = new(vertices[t.v1], vertices[t.v2], vertices[t.v3]);
+                if (triangles[triangleI].IsDegenerated)
+                    continue;
+                triangleIds[triangleI] = new(sectionI, section.map[localTriangleI]);
+                map[triangleI] = triangleI;
+                triangleI++;
+            }
+            triangleCounts[sectionI] = triangleI - baseMapIndices[sectionI];
+        }
 
-        splitStack ??= new(64);
+        int splitCount =
+            atomicCollisions.Sum(c => c.splits.Length) +
+            Math.Max(1, world.FindAllChildrenById(SectionId.PlaneSection, true).Count());
+        var rootPlane = world.FindChildById(SectionId.PlaneSection, false) as RWPlaneSection;
+        var rootAtomic = world.FindChildById(SectionId.AtomicSection, false);
+        var rootSection = rootPlane ?? rootAtomic ?? throw new InvalidDataException("RWWorld has no geometry");
+
+        // add dummy plane for root atomic worlds
+        rootPlane ??= new RWPlaneSection()
+        {
+            sectorType = RWPlaneSectionType.XPlane,
+            leftValue = float.PositiveInfinity,
+            rightValue = float.PositiveInfinity,
+            children = [rootAtomic, new RWString()]
+        };
+
+        var splits = new CollisionSplit[splitCount];
+        var splitI = 1;
+        var planeStack = new Stack<(RWPlaneSection, int)>();
+        planeStack.Push((rootPlane!, 0));
+        while (planeStack.TryPop(out var tuple))
+        {
+            var (plane, planeSplitI) = tuple;
+            var leftSection = plane.children[0];
+            var rightSection = plane.children[1];
+            var splitType = plane.sectorType switch
+            {
+                RWPlaneSectionType.XPlane => CollisionSectorType.X,
+                RWPlaneSectionType.YPlane => CollisionSectorType.Y,
+                RWPlaneSectionType.ZPlane => CollisionSectorType.Z,
+                _ => throw new InvalidDataException()
+            };
+
+            splits[planeSplitI] = new()
+            {
+                left = ProcessSubSection(splitType, plane.leftValue, leftSection),
+                right = ProcessSubSection(splitType, plane.rightValue, rightSection)
+            };
+        }
+
+        return new WorldCollider(box, new()
+        {
+            map = map,
+            splits = splits
+        }, triangles, triangleIds, atomicSections);
+
+        CollisionSector ProcessSubSection(CollisionSectorType splitType, float value, Section section) => section switch
+        {
+            RWPlaneSection plane => ProcessSubPlane(splitType, value, plane),
+            RWAtomicSection atomic => ProcessSubAtomic(splitType, value, atomic),
+            RWString => new() // dummy atomic
+            {
+                type = splitType,
+                value = value,
+                count = 0,
+                index = -1
+            },
+            _ => throw new ArgumentException("Unexpected section type", nameof(section))
+        };
+
+        CollisionSector ProcessSubPlane(CollisionSectorType splitType, float value, RWPlaneSection subPlane)
+        {
+            var subPlaneSplitI = splitI++;
+            planeStack.Push((subPlane, subPlaneSplitI));
+            return new()
+            {
+                type = splitType,
+                value = value,
+                count = RWCollision.SplitCount,
+                index = subPlaneSplitI
+            };
+        }
+
+        CollisionSector ProcessSubAtomic(CollisionSectorType splitType, float value, RWAtomicSection atomic)
+        {
+            var atomicId = atomicSections.IndexOf(atomic);
+            Debug.Assert(atomicId >= 0);
+            var subSplits = atomicCollisions[atomicId].splits;
+            if (subSplits.Length == 1 && float.IsInfinity(subSplits[0].left.value))
+            {
+                // naive collision where we can save the dummy split
+                return new()
+                {
+                    type = splitType,
+                    value = value,
+                    index = baseMapIndices[atomicId],
+                    count = triangleCounts[atomicId],
+                };
+            }
+
+            var startSplitI = splitI;
+            splitI += subSplits.Length;
+
+            subSplits.CopyTo(splits, startSplitI);
+            foreach (ref var subSplit in splits.AsSpan(startSplitI, subSplits.Length))
+            {
+                subSplit.left.index += subSplit.left.count == RWCollision.SplitCount
+                    ? startSplitI
+                    : baseMapIndices[atomicId];
+                subSplit.right.index += subSplit.right.count == RWCollision.SplitCount
+                    ? startSplitI
+                    : baseMapIndices[atomicId];
+            }
+
+            return new()
+            {
+                type = splitType,
+                value = value,
+                count = RWCollision.SplitCount,
+                index = startSplitI
+            };
+        }
     }
 
     public WorldTriangleInfo GetTriangleInfo(WorldTriangleId id) =>
         new(atomicSections[id.AtomicIdx], atomicSections[id.AtomicIdx].triangles[id.TriangleIdx]);
-
-    public override Raycast? Cast(Ray ray, float maxLength)
-    {
-        if (!Box.Intersects(ray.Start))
-        {
-            var coarse = ray.Cast(Box);
-            if (coarse == null || coarse.Value.Distance > maxLength)
-                return null;
-        }
-
-        var fine = RaycastSection(rootSection, ray, minDist: 0f, maxLength, prevHit: null);
-        if (fine is not null && fine.Value.Distance > maxLength)
-            return null;
-        return fine;
-    }
-
-    private Raycast? RaycastSection(Section section, Ray ray, float minDist, float maxDist, Raycast? prevHit)
-    {
-        if (minDist > maxDist)
-            return prevHit;
-
-        switch (section)
-        {
-            case RWAtomicSection atomic:
-                if (!atomicColliders.TryGetValue(atomic, out var atomicCollider))
-                    return prevHit;
-
-                var myHit = atomicCollider.Cast(ray, maxDist);
-                var isBetterHit = prevHit == null || (myHit != null && myHit.Value.Distance < prevHit.Value.Distance);
-                return isBetterHit
-                    ? myHit
-                    : prevHit;
-
-            case RWPlaneSection plane:
-                var compIndex = plane.sectorType.ToIndex();
-                var startValue = ray.Start.Component(compIndex);
-                var directionDot = ray.Direction.Component(compIndex);
-                var rightDist = ray.DistanceTo(plane.sectorType, plane.rightValue);
-                var leftDist = ray.DistanceTo(plane.sectorType, plane.leftValue);
-                var leftSection = plane.children[0];
-                var rightSection = plane.children[1];
-
-                Raycast? hit = prevHit;
-                if (directionDot < 0f)
-                {
-                    if (startValue >= plane.rightValue)
-                    {
-                        hit = RaycastSection(rightSection, ray, minDist, rightDist ?? maxDist, hit);
-                        float hitValue = hit?.Point.Component(compIndex) ?? float.MinValue;
-                        if (hitValue > plane.leftValue)
-                            return hit;
-                    }
-                    hit = RaycastSection(leftSection, ray, leftDist ?? minDist, maxDist, hit);
-                }
-                else
-                {
-                    if (startValue <= plane.leftValue)
-                    {
-                        hit = RaycastSection(leftSection, ray, minDist, leftDist ?? maxDist, hit);
-                        float hitValue = hit?.Point.Component(compIndex) ?? float.MaxValue;
-                        if (hitValue < plane.rightValue)
-                            return hit;
-                    }
-                    hit = RaycastSection(rightSection, ray, rightDist ?? minDist, maxDist, hit);
-                }
-                return hit;
-
-            default:
-                throw new InvalidDataException("Unexpected non-world section");
-        }
-    }
-
-    private static Stack<Section> splitStack = new();
-
-    protected override IEnumerable<Intersection> Intersections<T, TQueries>(T primitive)
-    {
-        var l = new List<Intersection>(32);
-        IntersectionsList<T, TQueries>(primitive, l);
-        return l;
-    }
-
-    public void IntersectionsList<T, TQueries>(in T primitive, List<Intersection> intersections)
-        where T : struct, IIntersectable
-        where TQueries : IIntersectionQueries<T>
-    {
-        //var splitStack = new Stack<Section>();
-        splitStack.Clear();
-        splitStack.Push(rootSection);
-        while (splitStack.TryPop(out var section))
-        {
-            switch (section)
-            {
-                case RWAtomicSection atomic when atomicColliders.TryGetValue(atomic, out var collider):
-                    collider.IntersectionsList<T, TQueries>(primitive, intersections);
-                    break;
-
-                case RWPlaneSection plane:
-                    int planeType = (int)plane.sectorType / 4;
-                    var leftSection = plane.children[0];
-                    var rightSection = plane.children[1];
-
-                    if (TQueries.SideOf(planeType, plane.rightValue, primitive) != PlaneIntersections.Outside)
-                        splitStack.Push(rightSection);
-                    if (TQueries.SideOf(planeType, plane.leftValue, primitive) != PlaneIntersections.Inside)
-                        splitStack.Push(leftSection);
-                    break;
-            }
-        }
-    }
 }
